elearning/src/main/java/com/ncnmo/aspire/elearning/service/UserService.java
package com.ncnmo.aspire.elearning.service;

import com.ncnmo.aspire.elearning.dto.UserRegistrationDTO;
import com.ncnmo.aspire.elearning.model.Role;
import com.ncnmo.aspire.elearning.model.User;
import com.ncnmo.aspire.elearning.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class UserService {

    @Autowired
    private UserRepository userRepository;

    @Autowired
    private PasswordEncoder passwordEncoder;

    public User registerUser(UserRegistrationDTO userRegistrationDTO) {
        return saveUser(userRegistrationDTO, Role.ROLE_USER);
    }

    public User registerAdmin(UserRegistrationDTO userRegistrationDTO) {
        return saveUser(userRegistrationDTO, Role.ROLE_ADMIN);
    }

    private User saveUser(UserRegistrationDTO userRegistrationDTO, Role role) {
        // Validate password and confirmPassword
        if (!userRegistrationDTO.getPassword().equals(userRegistrationDTO.getConfirmPassword())) {
            throw new IllegalArgumentException("Password and Confirm Password do not match");
        }

        User user = new User();
        user.setFirstName(userRegistrationDTO.getFirstName());
        user.setLastName(userRegistrationDTO.getLastName());
        user.setUsername(userRegistrationDTO.getUsername());
        user.setEmail(userRegistrationDTO.getEmail());
        user.setPassword(passwordEncoder.encode(userRegistrationDTO.getPassword()));
        user.setRole(role);

        return userRepository.save(user);
    }

    public User saveUser(User user) {
        return userRepository.save(user);
    }

    public String encodePassword(String rawPassword) {
        return passwordEncoder.encode(rawPassword);
    }

    public Optional<User> findByUsername(String username) {
        return userRepository.findByUsername(username);
    }

    public List<User> findAll() {
        return userRepository.findAll();
    }

    public void deleteUser(Long id) {
        userRepository.deleteById(id);
    }

    public Optional<User> findById(Long id) {
        return userRepository.findById(id);
    }
}
